requirements.txt
flash
python-dotenv
openai
